requirements.txt
django==1.4
django-tastypie
#django-jython
pytz
#requests
simplejson
mysql-python




